# internal/handler/group.go
package handler

import (
	"edusync/internal/model"
	"edusync/pkg/validator"

	"github.com/gin-gonic/gin"
)

// Create Group
// @Description Create Group
// @Summary Create Group
// @Tags Group
// @Accept json
// @Produce json
// @Param create body model.GroupCreateRequest true "Create Group"
// @Success 200 {object} model.BaseResponse
// @Failure 400 {object} model.BaseResponse
// @Failure 404 {object} model.BaseResponse
// @Failure 500 {object} model.BaseResponse
// @Router /api/groups [post]
// @Security ApiKeyAuth
func (h *Handler) createGroup(c *gin.Context) {
	var (
		err   error
		input model.GroupCreateRequest
	)

	if err = c.ShouldBindJSON(&input); err != nil {
		errorResponse(c, BadRequest, err)
		return
	}

	if err := validator.ValidatePayloads(input); err != nil {
		errorResponse(c, BadRequest, err)
		return
	}

	id, err := h.services.GroupService.GroupWriter.Create(input)
	if err != nil {
		fromError(c, err)
		return
	}

	successResponse(c, Created, id, nil)
}

// Get List Group
// @Description Get List Group
// @Summary Get List Group
// @Tags Group
// @Accept json
// @Produce json
// @Param pageSize query int64 true "pageSize" default(10)
// @Param page  query int64 true "page" default(1)
// @Success 200 {object} model.BaseResponse
// @Failure 400 {object} model.BaseResponse
// @Failure 404 {object} model.BaseResponse
// @Failure 500 {object} model.BaseResponse
// @Router /api/groups [get]
// @Security ApiKeyAuth
func (h *Handler) getListGroup(c *gin.Context) {
	pagination, err := listPagination(c)
	if err != nil {
		errorResponse(c, BadRequest, err)
		return
	}

	groups, err := h.services.GroupService.GroupReader.GetList(&pagination)
	if err != nil {
		fromError(c, err)
		return
	}

	successResponse(c, OK, groups, &pagination)
}

// Get Group By Id
// @Description Get Group By Id
// @Summary Get Group By Id
// @Tags Group
// @Accept json
// @Produce json
// @Param id path int64 true "Group Id"
// @Success 200 {object} model.BaseResponse
// @Failure 400 {object} model.BaseResponse
// @Failure 404 {object} model.BaseResponse
// @Failure 500 {object} model.BaseResponse
// @Router /api/groups/{id} [get]
// @Security ApiKeyAuth
func (h *Handler) getGroupById(c *gin.Context) {
	id, err := getNullInt64Param(c, idQuery)
	if err != nil {
		errorResponse(c, BadRequest, err)
		return
	}

	group, err := h.services.GroupService.GroupReader.GetById(id)
	if err != nil {
		fromError(c, err)
		return
	}

	successResponse(c, OK, group, nil)
}

// Update Group
// @Description Update Group
// @Summary Update Group
// @Tags Group
// @Accept json
// @Produce json
// @Param id path int64 true "Group Id"
// @Param update body model.GroupUpdateRequest true "Update Group"
// @Success 200 {object} model.BaseResponse
// @Failure 400 {object} model.BaseResponse
// @Failure 404 {object} model.BaseResponse
// @Failure 500 {object} model.BaseResponse
// @Router /api/groups/{id} [put]
// @Security ApiKeyAuth
func (h *Handler) updateGroup(c *gin.Context) {
	var input model.GroupUpdateRequest

	if err := c.ShouldBindJSON(&input); err != nil {
		errorResponse(c, BadRequest, err)
		return
	}

	if err := validator.ValidatePayloads(input); err != nil {
		errorResponse(c, BadRequest, err)
		return
	}

	id, err := getNullInt64Param(c, idQuery)
	if err != nil {
		errorResponse(c, BadRequest, err)
		return
	}
	input.Id = id

	err = h.services.GroupService.GroupWriter.Update(input)
	if err != nil {
		fromError(c, err)
		return
	}

	successResponse(c, OK, nil, nil)
}

// Delete Group
// @Description Delete Group
// @Summary Delete Group
// @Tags Group
// @Accept json
// @Produce json
// @Param id path int64 true "Group Id"
// @Success 200 {object} model.BaseResponse
// @Failure 400 {object} model.BaseResponse
// @Failure 404 {object} model.BaseResponse
// @Failure 500 {object} model.BaseResponse
// @Router /api/groups/{id} [delete]
// @Security ApiKeyAuth
func (h *Handler) deleteGroup(c *gin.Context) {
	id, err := getNullInt64Param(c, idQuery)
	if err != nil {
		errorResponse(c, BadRequest, err)
		return
	}

	err = h.services.GroupService.GroupWriter.Delete(id)
	if err != nil {
		fromError(c, err)
		return
	}

	successResponse(c, OK, nil, nil)
}